Updated with feedback - Validation for sqs parameter missing in xml config - Made SqsMessageHandlerWithQueueMessagingTemplateTests.java use the same tests as SqsMessageHandlertests by moving the tests to a common parent class: AbstractSqsMessageHandlerTest - Other minor fixes.
